PR20713: fix out of bounds access in iluGaussian
the 'rightmost pixel' case in iluBlurGaussian was the same as the
'leftmost pixel' case, which makes it refer to pixels that are off the
right-hand side of the image by 1 or 2 pixels.  In the final row they
are beyond the end of the image, and if the bytes just beyond the end of
the image are in an unmapped page, the program can segfault.

change the 'rightmost' case to use the pixels to the left of the pixel
being filtered instead.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>